pyproject.toml
[project]
name = "slopstopper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "google-genai>=0.3.0",
    "matplotlib>=3.10.8",
    "pandas>=2.2.3",
    "plotly>=5.24.1",
    "pydantic>=2.10.4",
    "python-dotenv>=1.0.1",
    "rich>=13.9.4",
    "sqlite-utils>=3.38",
    "streamlit>=1.41.1",
    "yt-dlp>=2024.12.13",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
pythonpath = ["."]
